Add reporting time zone to the bottom of the tab
