+++ Modules/Sources/JetpackStats/Screens/TrafficTabView.swift
@@ -17,6 +17,12 @@ struct TrafficTabView: View {
                 ForEach(viewModels, id: \.id) { viewModel in
                     makeItem(for: viewModel)
                 }
+
+                // Timezone info at the bottom with inset
+                TimezoneInfoView()
+                    .padding(.horizontal, Constants.step4)
+                    .padding(.top, Constants.step2)
+                    .padding(.bottom, Constants.step1)
             }
             .padding(.vertical, Constants.step2)
         }
